move banner inside advertisements folder
