+++ src/components/application/advertisements/banner/Banner.tsx
