+++ src/components/application/advertisements/banner/index.ts
