+++ src/components/templates/InfiniteScrollPageContent.tsx
@@ -4,7 +4,7 @@ import React, { useCallback, useEffect, useMemo } from "react";
 
 import { advertisementsFetcher } from "@/api/fetcher";
 import { CardsGrid } from "@/components/application/advertisements/cardsGrid";
-import { Banner } from "@/components/application/banner";
+import { Banner } from "@/components/application/advertisements/banner";
 import { LoadingSpinner } from "@/components/ui/loadingSpinner";
 import useInfiniteScroll from "@/hooks/useInfiniteScroll";
 import { Advertisement } from "@/types/api";
