+++ src/components/templates/PaginatedPage/ContentGenerator.tsx
@@ -1,5 +1,5 @@
 import { CardsGrid } from "@/components/application/advertisements/cardsGrid";
-import { Banner } from "@/components/application/banner";
+import { Banner } from "@/components/application/advertisements/banner";
 import { Advertisement } from "@/types/api";
 
 import {
